FETUS2026_Final_Baseline/requirements.txt
absl-py==2.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
docutils==0.21.2
einops==0.8.1
filelock==3.20.0
fsspec==2025.12.0
grpcio==1.76.0
h5py==3.15.1
id==1.5.0
idna==3.4
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monai==1.4.0
mpmath==1.3.0
networkx==3.4.2
nh3==0.2.21
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pillow==12.0.0
pip==25.3
protobuf==6.33.2
Pygments==2.19.1
PyWavelets==1.8.0
readme_renderer==44.0
requests==2.28.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
scikit-image==0.21.0
scikit-learn==1.7.2
scipy==1.15.3
setuptools==80.9.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
twine==6.1.0
typing_extensions==4.15.0
urllib3==1.26.13
Werkzeug==3.1.4
wheel==0.45.1
